Add files via upload
Fixes for constraints 2, 3, and 4. 
-Constraint 2: Added code to ensure that single arcs do not overlap with double arcs.
-Constraint 3: The constraint now takes the sum of all arcs of a given teacher class combination across all days.
-Constraint 4: The constraint now takes a sum of all arcs of a given teacher class combination for a given day. 

Now, Htc and Ltc are now inputted as matrices with the rows being the teacher and columns the classes. When inputting the numbers, ensure that the sum of the hours within the matrix for a given teacher is <= to the total hours available that week. For example, if there are 4 days and 4 periods per day, a single teacher can work a maximum of 16 hours. Thus these 16 hours need to be distributed across all classes, so, for two classes we may have 8 hours for class 1 and 8 for class to or 6 hours for class 1 and 10 for class 2. Any combination is suitable so long as the total sum is less than #Days*#Periods.