Added individual teacher timetables
IMPORTANT: You need to install the library "xlwt" for this to work. You may use pip install to do this. 

This updated file now also creates a .xls file which gives the class schedule for each teacher, which gives a cleaner view of the timetable. However, a bit of overview is lost - which is why the original timetable is also kept. 

These two are distinguished through the prefix of "Individual" and "Grouped" respectively.